Pulled generic modal creation function and associated pieces out to be used with other menus not specific to managing score periods
