+++ src/pieces/discussion/menu/staff/manageScorePeriods/ModalComponents.ts
@@ -1,7 +1,6 @@
 import { ActionRowBuilder, TextInputBuilder, TextInputStyle } from "discord.js";
 
 // MODAL BEHAVIOR CONSTANTS
-export const MODAL_EXPIRATION_TIME = 600_000; // 10 minutes
 export const DATE_STRING_FORMAT = "yyyy-MM-dd hh:mm:ss a";
 
 // MODAL NOTIFICATION CONSTANTS
